I/O

I/O

Input

Annotation files are read with readgbk(pathtofile). Currently this assumes that the file follows standard GenBank format.

GenomicAnnotations.readgbk — Method.

readgbk(filename)

Parse GenBank-formatted file filename, returning a Vector{Chromosome}.

source

Output

Annotations can be printed as with GenBank formatting using printgbk. In the REPL, instances of Gene are displayed as they would be in the annotation file.

GenomicAnnotations.printgbk — Function.

printgbk([io], chr)

Print chr in GenBank format.

source