I/O

Input

Annotation files are read with readgbk(input) or readgff(input). Currently these assume that the file follows either standard GenBank format, or GFF3. Any metadata in GFF3 files, apart from the header, is ignored.

GenomicAnnotations.readgbk — Function

readgbk(input, G::Type = Gene; gunzip = false)

Parse GenBank-formatted file, returning a Vector{Chromosome}. input can be a file path or an IOStream. File names ending in ".gz" are assumed to be gzipped and are decompressed. Setting gunzip to true forces this behaviour. The type of AbstractGene to be used can be specified with G, though currently the only option is Gene.

GenomicAnnotations.readgff — Function

readgff(input, G::Type = Gene; gunzip = false)

Parse GFF3-formatted file, returning a Vector{Chromosome}. input can be a file path or an IOStream. File names ending in ".gz" are assumed to be gzipped and are decompressed. Setting gunzip to true forces this behaviour. The type of AbstractGene to be used can be specified with G, though currently the only option is Gene.

Output

Annotations can be printed with GenBank formatting using printgbk, and as GFF3 with printgff. Headers are not automatically converted between formats; printgff only prints the header of the first Chromosome, and only if it starts with a #, while printgbk prints a default header if the stored one starts with #.

GenomicAnnotations.printgbk — Function

printgbk(io::IO, chr)
printgbk(path::AbstractString, chr)

Print chr in GenBank format.

GenomicAnnotations.printgff — Function

printgff(io::IO, chr)
printgff(path::AbstractString, chr)

Print chr in GFF3 format.

In the REPL, instances of Gene are displayed as they would be in the annotation file.